requirements.txt
flask>=3.0
requests>=2.31
gunicorn>=21.2
google-genai>=1.0
